Use an abstract schema and support PostgreSQL
This switches to the abstract schema specification. To (hopefully)
simplify the maintenance of the abstract schema going forward, the two
separate SQL file definitions of the geo_tags table based on the
backend have been replaced with one table definition and one (optional)
patch.

To facilitate migration from databases that predate this change, we
check if the database has one of the fields introduced by the patch
to see if it's necessary. This is necessary because pre-existing
installations would fail to apply the patch because it's (effectively)
already applied.

The one thing lost in this change is the _bin collation on the gt_name
column. This is deemed acceptable because it only affects collation and
is automatically dropped in SQLite. Pre-existing installations will
keep the _bin collation type.

Some of the code paths triggered non-fully-portable code paths (broken
on PostgreSQL) in MediaWiki core, such as booleans not being portably
defined as integers and explicit NULLs being provided to autoincrement
columns being prohibited. Those code paths have been altered to only
create portable SQL.

Bug: T268550